Move terminal-kind command handling onto TerminalSurfaceManager
The terminal-kind command handler and its stop helper lived as private
funcs on the neutral WorktreeSurfaceManager, so each new kind would have
grown the shared file by a handler's worth of kind code. They now live
on TerminalSurfaceManager, a kind-level coordinator the core routes to
without interpreting the payload; a future kind adds a sibling manager
and one dispatch arm.

One coordinator serves every worktree: commands arrive addressed by
Worktree, and per-worktree resolution stays behind the core's
state(for:) / stateIfExists(for:) split so the stop arms keep acting on
existing state without minting any. forceEmitProjection widens to
internal for the stale-mirror reconcile path.
